@@ -37,6 +37,6 @@
     "karma-coverage": "~2.2.1",
     "karma-jasmine": "~5.1.0",
     "karma-jasmine-html-reporter": "~2.1.0",
-    "typescript": "~5.1.6"
+    "typescript": "~5.2.2"
   }
 }
