@@ -9,11 +9,10 @@ import {HeaderComponent} from './core/component/header/header.component';
 import {LoginFormComponent} from './login-form/login-form.component';
 import {ContentComponent} from './content/content.component';
 
-import {AxiosService} from './axios.service';
 import { HomepageComponent } from './homepage/homepage.component';
 import {LoginComponent} from "./core/component/login/login.component";
 import {FooterComponent} from "./core/component/footer/footer.component";
-import {HttpClient, HttpClientModule} from "@angular/common/http";
+import {HttpClientModule} from "@angular/common/http";
 
 @NgModule({
   declarations: [
@@ -32,7 +31,6 @@ import {HttpClient, HttpClientModule} from "@angular/common/http";
     FormsModule,
     HttpClientModule
   ],
-  providers: [AxiosService],
   bootstrap: [AppComponent]
 })
 export class AppModule {
