@@ -1,7 +1,5 @@
 import { Component } from '@angular/core';
 import {AuthService} from "../../service/security/auth.service";
-import {BaseFormComponent} from "../base-form/base-form.component";
-import {UntypedFormBuilder, Validators} from '@angular/forms';
 import {Router} from "@angular/router";
 
 @Component({
@@ -21,20 +19,8 @@ export class LoginComponent  {
   submit() {
       if(this.login && this.password) {
         this.authService.login(this.login, this.password).subscribe({
-          next: () => this.router.navigate(['main-page'])
+          next: () => this.router.navigate(['main-page']),
         })
       }
   }
-
-  // protected override setupForm(): any[] {
-  //   return [
-  //     {
-  //       login: ['', [Validators.required]],
-  //       password: ['', Validators.required]
-  //     }
-  //   ]
-  // }
-
-
-
 }
